requirements.txt
PyQt6>=6.4.0
pywinpty>=2.0.0; sys_platform == "win32"
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0
